+++ src/com/muflihun/residue/ResidueConnectTask.java
@@ -1,3 +1,21 @@
+/**
+ * ResidueConnectTask.java
+ * <p>
+ * Connect activity for android apps
+ * Part of Residue Java client library
+ * <p>
+ * Copyright (C) 2017-present Muflihun Labs
+ * <p>
+ * https://muflihun.com
+ * https://muflihun.github.io/residue
+ * https://github.com/muflihun/residue-java
+ * <p>
+ * See https://github.com/muflihun/residue-java/blob/master/LICENSE
+ * for licensing information
+ * <p>
+ * Author: @abumusamq
+ */
+
 package com.muflihun.residue;
 
 import android.os.AsyncTask;