+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [1.1.0] - 09-02-2018
+### Updates
+- License information update
+
+### Fixed
+- Fixed crash after license information from server is no longer available
+- `ResiduePrintStream` class completed
+
 ## [1.0.1]
 ### Updates
 - Compatibility with residue v1.2.0